requirements.txt
blobfile
mpi4py
Pillow
numpy
torch==1.13
torchvision
seaborn
tqdm
